Added REPL engine
Added a REPL engine to read input and execute commands.

The commands will be the hook into calling the application layer.
Cucumber is using the REPL class the communicate with the SocialNetwork
Ofcourse, the cucumber tests are now failing because there are no
commands yet.

The REPL engine has a quit command: \q will break the loop and will exit
the application.