Created all controllers, and they call the UseCase objects.
Controllers are created to first check if they support a given command.
All commands supporting the given input will be executed. Currently all
controllers are implemented and they all call a UseCase object from the
application layer.

The Controllers are using regex patterns to determine if they support
the input. If so, they use the regex to extract the variables from the
inputLine. This is a bit brittle, but it's the easiest way to parse user
input.

Had to fight a bit with regex obviously because the read timeline was
matching a bit too much :) damned regex!