Splitting Request Parsing and Controllers
The controllers contained logic to parse the request from the CLI. This
logic has been moved to a RequestFactory. This factory will return a
Request object for each type of request.
The RequestDispatcher is responsible for dispatching the requests to the
correct Controller. It uses a visitor pattern with double dispatch to
forward the requests.

One downside of this approach is that we now have fixed list of
Controllers which are supported. If you want to add a Controller for a
new usecase, you are now forced to adapt the Visitor class (The
RequestDispatcher) and then add the Controller.
The previous implementation did not have this shortcoming, You would
simply add a Controller to the list, and that's it.